refactor(757): compat scaffold + latent-import fixes ahead of the namespace split
First step of the #757 workflow-namespace migration (see plan): structural,
no behavior change.

- crossval: import `_as_topic_word` from `.coherence` (its real home) instead of
  `.validation`, which only re-exported it. The current code works only because
  validation re-exports the symbol; the upcoming validation.py split would break
  that import.
- Move the cross-concern `_ref_corpus` helper from validation.py to coherence.py
  (next to `_as_topic_word`); validation.py re-exports it, so `diagnostics`,
  `search_k`, and any `from .validation import _ref_corpus` are unaffected. This
  gives the helper a home that both the future `select` and `evaluate` modules can
  import without a cycle.
- Verified the analysis <-> validation access is already cycle-safe (analysis binds
  the validation module object and defers attribute access to call time;
  validation imports analysis only inside functions), so the split forms no
  import cycle.

Verification: full suite 4959 passed / 38 skipped.
